// client/src/app/_models/address.ts
export class Address {
  Id: number = 0;
  Name: string = '';
  Address1: string = '';
  Address2: string = '';
  City: string = '';
  County: string = '';
  Postcode: string = '';
  CountryCode: string = '';
  Telephone: string = '';
  Email: string = '';
  Fax: string = '';
  UserId: string = '';
}